+++ packages/uxpin-merge-cli/src/program/command/Command.ts
@@ -1,6 +1,8 @@
 export enum Command {
   DUMP = 'dump',
   EXPERIMENT = 'experiment',
+  INIT = 'init',
+  GENERATE_PRESETS = 'generate-presets',
   PUSH = 'push',
   SERVER = 'server',
   SUMMARY = 'summary',