+++ packages/uxpin-merge-cli/src/resources/Button/presets/0-default.jsx
@@ -1,0 +1,10 @@
+import * as React from "react";
+import Button from "../Button";
+
+export default (
+  <Button
+    uxpId="button"
+    disabled={false}
+    label="Button">
+  </Button>
+);