+++ packages/uxpin-merge-cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [2.7.3] - 2021-03-01
+
+### Added
+- Added an init command which adds the default files required by merge
+- Added a generate-presets command which reads the config file and auto generates preset files
+
 ## [2.7.2] - 2020-12-15
 
 ### Fixed